Refactor ok-api's packages
