+++ ok-api/src/main/java/polis/ok/api/OkAuthorizator.java
@@ -6,6 +6,7 @@ import org.json.JSONException;
 import org.json.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import polis.ok.api.exceptions.OkApiException;
 
 import java.io.IOException;
 import java.net.URI;
